Creacion endpoint Rol, usuarioEmpleado, cliente, persona y compras

+++ backend_veterinaria/src/main/java/com/veterinaria/backend_veterinaria/repository/MascotaRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.veterinaria.backend_veterinaria.models.Mascota;
 
-public interface MascotaRepository extends JpaRepository<Mascota, Long> {}  
+public interface MascotaRepository extends JpaRepository<Mascota, Integer> {}
